requirements.txt
elasticsearch
pandas
python-dotenv
flask
requests
bs4
